requirements.txt
django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
pillow==10.0.0
razorpay==1.3.0
django-cors-headers==4.2.0
python-decouple==3.8